+++ .github/scripts/release.sh
@@ -5,6 +5,7 @@ set -ex
 OLD_VERSION="$1"
 NEW_VERSION="$2"
 # Use SED_BINARY from environment if set, otherwise default to 'sed'
+# On MacOs - install gnu-sed with brew
 SED_BINARY=${SED_BINARY:-sed}
 
 ${SED_BINARY} -i "s/$OLD_VERSION/$NEW_VERSION/g" Cargo.toml